add queries to Meta triplestore
The process now features the possibility of querying OC Meta SPARQL endpoint in order to check if an external ID "exists", i.e. is already registered somewhere.
If IdExistence.use_meta_endpoint is set to True, the program first checks the presence of an ID in Meta. Then, if it is there, it just holds it as existing (check_id_existence returns True). If it is not there, it looks for it in the appropriate external database via API. If IdExistence.use_meta_endpoint is set to False, only external services are used.
In this way, the process is faster with IDs that already exist in Meta (since querying Meta SPARQL endpoint is faster than querying external APIs) but slower when querying IDs that are not already present in Meta (because if an ID is not found in Meta, an external API is called anyway in order to verify its existence).