update Wellformedness.get_missing_values
venue is always mandatory if volume or issue are specified. This applies for every type of BR and regardless of the presence of an ID.
N.B.: the table in docs/mandatory_fields.csv also implies (in the last two lines) which types of require the venue if the volume/issue is specified and the ID is specified too. These types of resources are not taken into cosideration in the validator implementation, allowing for stricter behaviour.