bugfix + include temp: and local: IDs
Bugfix: in main.py:873 the message under "if self.strict_sequentiality" was always printed regardless of the presence of previously detected errors -> now it is only printed when  self.strict_sequentiality is set to True AND errors have already been found.

Inclusion of :temp and :local as supported IDs: so far, they are only intended to be used as placehoders (internal temporary IDs) when a resource that is represented as first class entity lacks external PIDs, i.e. temp: or :local can only be used in the "id" field of META-CSV and "citing_id"/"cited_id" fields of CITS-CSV.

Other:
- messages.yaml: removed extra whitespace characters
- make message of InvalidTableError clearer for end user