bugfix: local IDs
Local IDs (with scheme "local" or "temp") were treated like public persistent identifiers for the purpose of identifying an entity. This has been fixed, as they should only be used to map entities between the metadata table and the citation table. --> Now it is mandatory to specify the required metadata for entities that have no other identifier than temp or local.